+++ tests/Behat/Page/Admin/Product/ShowPageInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\MangoSylius\ExtendedChannelsPlugin\Behat\Page\Admin\Product;
+
+use FriendsOfBehat\PageObjectExtension\Page\SymfonyPageInterface;
+
+interface ShowPageInterface extends SymfonyPageInterface
+{
+	public function duplicateProduct(): void;
+
+	public function getCodeValue(): string;
+}